revert renaming of pyControl folder
