+++ devices/poke.py
@@ -1,4 +1,4 @@
-from pycontrol_mcu.hardware import Digital_output, Digital_input
+from pyControl.hardware import Digital_output, Digital_input
 
 
 class Poke:
