+++ devices/stepper_motor.py
@@ -1,4 +1,4 @@
-from pycontrol_mcu.hardware import Digital_output
+from pyControl.hardware import Digital_output
 
 
 class Stepper_motor:
