+++ source/pyControl/readme.txt
@@ -1,0 +1,1 @@
+This folder contains files that run on the pyboard microcontroller
