+++ source/pyControl/utility.py
