--- source/pycontrol_mcu/__init__.py
@@ -1,4 +1,0 @@
-import pycontrol_mcu.framework as fw
-import pycontrol_mcu.utility as ut
-import pycontrol_mcu.hardware as hw
-import pycontrol_mcu.state_machine as sm
