+++ source/tests/board_tests/five_poke_test.py
@@ -6,7 +6,7 @@
 # Poke in each port of the 5 poke in turn, the light on the poke will illuminate
 # after you poke it.
 
-from pycontrol_mcu.utility import *
+from pyControl.utility import *
 from devices import Breakout_1_2, Five_poke
 
 # Define hardware
