+++ tasks/example/button.py
@@ -2,7 +2,7 @@
 # second when the usr pushbutton on the pyboard is pressed three times.
 # Does not require any hardware except a micropython board.
 
-from pycontrol_mcu.utility import *
+from pyControl.utility import *
 from devices import *
 
 # Define hardware
